+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 0.15.0
 
 ### Enhancements
+- **Annual Changes**
+  - Training grant administrators now receive an email when trainees update
+    their annual profile information
 - **Gem Changes**
   - Updated to jquery-rails 4.2.1